requirements.txt
python-pptx>=0.6.21
rapidfuzz>=3.0.0
openpyxl>=3.1.0
tqdm>=4.65.0
anthropic>=0.25.0
python-dotenv>=1.0.0
